Setter max poll interval til 15m som INTEGER

+++ app/src/main/kotlin/no/nav/paw/migrering/app/konfigurasjon/KafkaKonfigurasjon.kt
@@ -42,7 +42,7 @@ val KafkaKonfigurasjon.properties
         ProducerConfig.CLIENT_ID_CONFIG to klientKonfigurasjon.produsentKlientId,
         ConsumerConfig.AUTO_OFFSET_RESET_CONFIG to "earliest",
         ConsumerConfig.MAX_POLL_RECORDS_CONFIG to klientKonfigurasjon.maksHentetPerKall,
-        ConsumerConfig.MAX_POLL_INTERVAL_MS_CONFIG to Duration.ofMinutes(15).toMillis(),
+        ConsumerConfig.MAX_POLL_INTERVAL_MS_CONFIG to Duration.ofMinutes(15).toMillis().toInt(),
         CommonClientConfigs.BOOTSTRAP_SERVERS_CONFIG to serverKonfigurasjon.kafkaBrokers
     ) + if (serverKonfigurasjon.autentisering.equals("SSL", true)) {
         mapOf(
